+++ crypto/hd/fundraiser_test.go
